UI Update and Jury Fix
Integrated the latest version of Matt's UI (And updated Clue List and Clue Viewer to match).

Fixed some issues with the UI getting mislinked and throwing errors when loading the Jury scene.
Made sure that the Camera is not part of DontDestroyOnLoad.
Made sure that the that the UI itself is on DontDestroyOnLoad.