Fixed textbox issue
Added CanvasGroups to the dialogue and HUD canvases, and added two functions to CanvasManager: DisableCanvas and LoadCanvas. Note how their names are capitalized compared to loadCanvas and disableCanvas.

These new functions don't disable/enable the canvases, but they change the properties of the CanvasGroups to simulate disabling and enabling, without messing up the textbox's Close method.