Made nametag movable
Changed DialogueUITest so that it accepts this new yarn command, the syntax of which is:

<<NameTagMove|arg>>

Arg has to be either leftEdge or rightEdge, otherwise there'll be an error.

With this, we can have the captain's nametag be to the right, and everyone else's nametags be to the left.  Those writing the jsons will need to apply this new command to every textbox.