Testimony now pops up
Added two functions to CanvasManager. Changed the name of crewManger.cs to CrewManager.cs.

Had the singleton in CrewManager be set in CrewManger's Awake instead of Start. Avoids race condition problems, like what was mentioned in Level Building.

Changed a couple of the UI buttons so that they use the ShowCanvas and HideCanvas functions on the testimony canvas.

The Testimony Canvas is enabled by default, with just its CanvasGroup making it invisible.

Added TestimonyManager script to the Testimony Canvas.