Refined the testimony entries
They now only expand when there is enough text in them to justify it.

I also made it so that the scrollbar only shows up when there is enough stuff in the testimony window to warrant using it.

To fix a few issues, I had to add a ContentHeightFitter and ScrollbarSizer to make sure that the srollbar works properly. There is actually one slight glitch with the scrollbar, still, but we can worry about that later.

Please only merge these scripts into master:
ScrollbarEnabler
ContentHeightFitter
ScrollbarSizer
TestimonyEntry
CrewManager
TestimonyManager

And also merge the TestimonyEntry prefab. Because of my scene setup, there will probably be some issues with the scene in master, but when those pop up, please let me know.