Reimplemented the testimony stuff
Made a new scene called "Final_Ship (TestimonyStuff)" and managed to make things work in there. I added the scripts from my previous build, and changed a couple of things.

I added components to the Testimony Entry's Text, the Tweet_Groups object, and the Scrollbar_Vertical object.

I changed the scroll rect in the ScrollView object, too; I made it so that the Content object is Tweet_Groups, not Content (since it was actually Tweet_Groups that held the testimony).